refactor: Consolidate type definitions in types.py
